+++ bt/app/config.cpp
@@ -21,12 +21,9 @@ namespace bt {
     const string LogRuleHitsKey = "log_rule_hits";
     const string PersistPopularityKey = "persist_popularity";
     const string ShowHiddenBrowsersKey = "browsers_show_hidden";
+    const string ClearURLsEnabledKey = "clearurls_enabled";
 
-    common::config cfg{ 
-        fs::exists(fs::path{fss::get_current_dir()} / PortableMarkerName)
-            ? (fs::path{fss::get_current_dir()} / FileName).string()
-            : (fs::path{win32::shell::get_local_app_data_path()} / APP_SHORT_NAME / FileName).string()
-    };
+    common::config cfg{config::get_data_file_path(FileName)};
 
     config config::i;
 
@@ -46,6 +43,12 @@ namespace bt {
 
     std::string config::get_theme() {
         return cfg.get_value("theme");
+    }
+
+    std::string config::get_data_file_path(const std::string& name) {
+        return fs::exists(fs::path{fss::get_current_dir()} / PortableMarkerName)
+            ? (fs::path{fss::get_current_dir()} / name).string()
+            : (fs::path{win32::shell::get_local_app_data_path()} / APP_SHORT_NAME / name).string();
     }
 
     void config::ensure_instance_id() {
@@ -195,6 +198,15 @@ namespace bt {
 
     void config::set_show_hidden_browsers(bool show) {
         cfg.set_bool_value(ShowHiddenBrowsersKey, show);
+        cfg.commit();
+    }
+
+    bool config::get_clearurls_enabled() {
+        return cfg.get_bool_value(ClearURLsEnabledKey, false);
+    }
+
+    void config::set_clearurls_enabled(bool enabled) {
+        cfg.set_bool_value(ClearURLsEnabledKey, enabled);
         cfg.commit();
     }