+++ docs/release-notes.md
@@ -1,6 +1,16 @@
+## 3.5.0
+
+- todo: ClearURLs
+- todo: bug: "check for updates" was opening home page instead of GitHub releases.
+
+### New
+
+- Tray menu has "URL Tester" shortcut.
+- "buy me a coffee" link removed from the tray icon and status bar, because no one is donating.
+
 ## 3.4.2
 
-Bug fixed - picker was always defaulting to "chomeless window". Thanks to @danarnold in #20.
+Bug fixed - picker was always defaulting to "chrome-less window". Thanks to @danarnold in #20.
 
 ## 3.4.1